Move connection configuration models to a separate file
